% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_to_gfm.r
\name{rmd_to_gfm}
\alias{rmd_to_gfm}
\title{Rmd to GitHub flavored markdown for Jekyll Website}
\usage{
rmd_to_gfm(filename, pdf = FALSE)
}
\arguments{
\item{filename}{Name of Rmd file}

\item{pdf}{TRUE/FALSE  Do you want a pdf made?  If TRUE, make sure to add info to the yaml of Rmd}
}
\value{
None
}
\description{
Runs and Rmd file through knit.  Uses here, so you can run from 
anyplace in your project.  You don't need to worry about setting the 
working directory.  Will put the files in the right places.
}
\examples{
rmd_to_gfm("2016-5-19-FI-recursion-2.Rmd")
}
